Fixed issues from 62-68 And moved a file from body due to it no supposed to be there in in suspension instead
Merge remote-tracking branch 'origin/master'

# Conflicts:
#	edu/jalc/automobile/onlinebuilder/components/Driver.java
#	edu/jalc/automobile/onlinebuilder/components/README.md
#	edu/jalc/automobile/onlinebuilder/components/Suspension.java
#	edu/jalc/automobile/onlinebuilder/components/economy/test
#	edu/jalc/automobile/onlinebuilder/components/luxury/test
#	edu/jalc/automobile/onlinebuilder/components/sport/Sport.java
#	edu/jalc/automobile/onlinebuilder/components/sport/SportAdjustable.java
#	edu/jalc/automobile/onlinebuilder/components/sport/SportAdjustableTest.java
#	edu/jalc/automobile/onlinebuilder/components/sport/SportLowered.java
#	edu/jalc/automobile/onlinebuilder/components/sport/SportLoweredTest.java
#	edu/jalc/automobile/onlinebuilder/components/sport/SportRaised.java
#	edu/jalc/automobile/onlinebuilder/components/sport/SportRaisedTest.java
#	edu/jalc/automobile/onlinebuilder/components/towing/Towing.java
#	edu/jalc/automobile/onlinebuilder/components/towing/TowingHeavyDuty.java
#	edu/jalc/automobile/onlinebuilder/components/towing/TowingLightDuty.java
#	edu/jalc/automobile/onlinebuilder/components/towing/TowingSuperDuty.java
#	edu/jalc/automobile/onlinebuilder/components/towing/test
#	edu/jalc/automobile/parts/AdjustableShocks.java
#	edu/jalc/automobile/parts/AdjustableShocksTest.java
#	edu/jalc/automobile/parts/AllTerrainTire.java
#	edu/jalc/automobile/parts/ExtendedShocks.java
#	edu/jalc/automobile/parts/ExtendedSpring.java
#	edu/jalc/automobile/parts/ExtendedSpringTest.java
#	edu/jalc/automobile/parts/ForgedWheel.java
#	edu/jalc/automobile/parts/ForgedWheelTest.java
#	edu/jalc/automobile/parts/HeavyShock.java
#	edu/jalc/automobile/parts/HeavySpring.java
#	edu/jalc/automobile/parts/LightShock.java
#	edu/jalc/automobile/parts/LightSpring.java
#	edu/jalc/automobile/parts/MediumSpring.java
#	edu/jalc/automobile/parts/MediumSpringTest.java
#	edu/jalc/automobile/parts/README.md
#	edu/jalc/automobile/parts/Shock.java
#	edu/jalc/automobile/parts/ShortShocks.java
#	edu/jalc/automobile/parts/ShortSpring.java
#	edu/jalc/automobile/parts/ShortSpringTest.java
#	edu/jalc/automobile/parts/SportTire.java
#	edu/jalc/automobile/parts/SportTireTest.java
#	edu/jalc/automobile/parts/Spring.java
#	edu/jalc/automobile/parts/SteelWheel.java
#	edu/jalc/automobile/parts/SuperShock.java
#	edu/jalc/automobile/parts/SuperSpring.java
#	edu/jalc/automobile/parts/Tire.java
#	edu/jalc/automobile/parts/Wheel.java
#	src/edu/jalc/automobile/onlinebuilder/components/sport/SportAdjustable.java
#	src/edu/jalc/automobile/onlinebuilder/components/sport/SportLoweredTest.java
#	src/edu/jalc/automobile/onlinebuilder/components/suspension/sport/SportAdjustable.java
#	src/edu/jalc/automobile/onlinebuilder/components/suspension/sport/SportLoweredTest.java
#	src/edu/jalc/automobile/parts/AdjustableShocksTest.java
#	src/edu/jalc/automobile/parts/ExtendedSpringTest.java
#	src/edu/jalc/automobile/parts/ForgedWheelTest.java
#	src/edu/jalc/automobile/parts/MediumSpringTest.java
#	src/edu/jalc/automobile/parts/ShortSpringTest.java
#	src/edu/jalc/automobile/parts/SportTireTest.java
#	src/edu/jalc/automobile/parts/SuperShock.java
#	src/edu/jalc/automobile/parts/suspension/AdjustableShocksTest.java
#	src/edu/jalc/automobile/parts/suspension/ExtendedSpringTest.java
#	src/edu/jalc/automobile/parts/suspension/ForgedWheelTest.java
#	src/edu/jalc/automobile/parts/suspension/MediumSpringTest.java
#	src/edu/jalc/automobile/parts/suspension/ShortSpringTest.java
#	src/edu/jalc/automobile/parts/suspension/SportTireTest.java
#	src/edu/jalc/automobile/parts/suspension/SuperShock.java
