--- edu/jalc/automobile/onlinebuilder/components/README.md
@@ -1,2 +1,0 @@
-# Component
-Components used in creating automobiles
