--- edu/jalc/automobile/onlinebuilder/components/Suspension.java
@@ -1,5 +1,0 @@
-package edu.jalc.automobile.onlinebuilder.components;
-
-public abstract class Suspension{
-
-}
