--- edu/jalc/automobile/onlinebuilder/components/economy/test
@@ -1,1 +1,0 @@
-test1
