--- edu/jalc/automobile/onlinebuilder/components/luxury/test
@@ -1,1 +1,0 @@
-test
