--- edu/jalc/automobile/onlinebuilder/components/sport/Sport.java
@@ -1,7 +1,0 @@
-package edu.jalc.automobile.onlinebuilder.components.sport;
-
-import edu.jalc.automobile.onlinebuilder.components.*;
-
-abstract class Sport extends Suspension{
-
-}
