--- edu/jalc/automobile/onlinebuilder/components/sport/SportRaisedTest.java
@@ -1,28 +1,0 @@
-package edu.jalc.automobile.onlinebuilder.components.sport;
-
-import edu.jalc.automobile.parts.*;
-
-
-public class SportRaisedTest {
-
-     public void testGetShock(){
-        System.out.println("Testing SportRaised::getShock");
-        ExtendedShocks shock = new ExtendedShocks();
-        SportRaised sportRaised = new SportRaised(shock, null);
-        assert(shock == sportRaised.getShock());
-     }
-
-     public void testGetSpring(){
-        System.out.println("Testing SportRaised::getSpring");
-        ExtendedSpring spring = new ExtendedSpring(10);
-        SportRaised sportRaised = new SportRaised(null, spring);
-        assert(spring == sportRaised.getSpring());
-     }
-
-     public static void main(String[] args){
-        SportRaisedTest test = new SportRaisedTest();
-        test.testGetShock();
-        test.testGetSpring();
-        System.out.println("All Tests for SportRaised Passed");
-     }
-}
