--- edu/jalc/automobile/onlinebuilder/components/towing/Towing.java
@@ -1,7 +1,0 @@
-package edu.jalc.automobile.onlinebuilder.components.towing;
-
-import edu.jalc.automobile.onlinebuilder.components.*;
-
-abstract class Towing extends Suspension{
-
-}
