--- edu/jalc/automobile/onlinebuilder/components/towing/TowingLightDuty.java
@@ -1,32 +1,0 @@
-package edu.jalc.automobile.onlinebuilder.components.towing;
-
-import edu.jalc.automobile.parts.*;
-
-public class TowingLightDuty extends Towing{
-
-   private Shock shock;
-   private Spring spring;
-   
-   private TowingLightDuty(){
-      this.shock = null;
-      this.spring = null;
-   }
-   
-   public TowingLightDuty(Shock shock, Spring spring){
-      this.shock = shock;
-      this.spring = spring;
-   }
-   
-   public Shock getShock(){
-      return shock;
-   }
-
-   public Spring getSpring(){
-      return spring;
-   }
-   
-   public String toString(){
-      return shock.toString() + ", " + spring.toString();
-   }
-   
-}
