--- edu/jalc/automobile/onlinebuilder/components/towing/test
@@ -1,1 +1,0 @@
-test4
