--- edu/jalc/automobile/parts/AdjustableShocks.java
@@ -1,22 +1,0 @@
-package edu.jalc.automobile.parts;
-
-public class AdjustableShocks extends Shock{
-
-   private double height;
-
-   private AdjustableShocks(){
-      this.height = 0;
-   }
-
-   public AdjustableShocks(double height){
-      this.height = height;
-   }
-
-   public double getHeight(){
-      return height;
-   }
-
-   public String toString(){
-      return height + " inch shocks";
-   }
-}
