--- edu/jalc/automobile/parts/AllTerrainTire.java
@@ -1,29 +1,0 @@
-package edu.jalc.automobile.parts;
-
-public class AllTerrainTire extends Tire{
-
-   private double height;
-   private double width;
-   
-   private AllTerrainTire(){
-      this.height = 0;
-      this.width = 0;
-   }
-   
-   public AllTerrainTire(double height, double width){
-      this.height = height;
-      this.width = width;
-   }
-   
-   public double getHeight(){
-      return height;
-   }
-   
-   public double getWidth(){
-      return width;
-   }
-   
-   public String toString(){
-      return "All Terrain Tire";
-   }
-}
