--- edu/jalc/automobile/parts/ExtendedShocks.java
@@ -1,8 +1,0 @@
-package edu.jalc.automobile.parts;
-
-public class ExtendedShocks extends Shock{
-
-   public String toString(){
-      return "12 inch shocks";
-   }
-}
