--- edu/jalc/automobile/parts/HeavyShock.java
@@ -1,8 +1,0 @@
-package edu.jalc.automobile.parts;
-
-public class HeavyShock extends Shock{
-
-   public String toString(){
-      return "Heavy Shock";
-   }
-}
