--- edu/jalc/automobile/parts/LightShock.java
@@ -1,8 +1,0 @@
-package edu.jalc.automobile.parts;
-
-public class LightShock extends Shock{
-
-   public String toString(){
-      return "Light Shock";
-   }
-}
