--- edu/jalc/automobile/parts/README.md
@@ -1,2 +1,0 @@
-# Parts
-Library for automobile parts
