--- edu/jalc/automobile/parts/ShortSpring.java
@@ -1,18 +1,0 @@
-package edu.jalc.automobile.parts;
-
-public class ShortSpring extends Spring{
-
-   public ShortSpring(double height) {
-      this.height = height;
-   }
-
-   private double height = 2;
-
-   public double getHeight(){
-      return height;
-   }
-
-   public String toString(){
-      return height+ " inch Coil Spring";
-   }
-}
