--- edu/jalc/automobile/parts/SportTire.java
@@ -1,29 +1,0 @@
-package edu.jalc.automobile.parts;
-
-public class SportTire extends Tire{
-
-   private double height;
-   private double width;
-
-   private SportTire(){
-      this.height = 0;
-      this.width = 0;
-   }
-
-   public SportTire(double height, double width){
-      this.height = height;
-      this.width = width;
-   }
-
-   public double getHeight(){
-      return height;
-   }
-
-   public double getWidth(){
-      return width;
-   }
-
-   public String toString(){
-      return "Sport Tire";
-   }
-}
