--- edu/jalc/automobile/parts/Spring.java
@@ -1,5 +1,0 @@
-package edu.jalc.automobile.parts;
-
-public abstract class Spring{
-
-}
