--- edu/jalc/automobile/parts/SuperShock.java
@@ -1,8 +1,0 @@
-package edu.jalc.automobile.parts;
-
-public class SuperShock extends Shock{
-
-   public String toString(){
-      return "Super Shock";
-   }
-}
