--- edu/jalc/automobile/parts/SuperSpring.java
@@ -1,23 +1,0 @@
-package edu.jalc.automobile.parts;
-
-public class SuperSpring extends Spring{
-
-   private double height;
-   
-   private SuperSpring(){
-      this.height = 0;
-   }
-   
-   public SuperSpring(double height){
-      this.height = height;
-   }
-   
-   public double getHeight(){
-      return height;
-   }
-   
-   public String toString(){
-      return "Super Leaf Spring";
-   }
-}
-
