+++ src/edu/jalc/automobile/parts/suspension/SuperShock.java
@@ -1,0 +1,9 @@
+package edu.jalc.automobile.parts.suspension;
+
+public class SuperShock extends Shock{
+
+   public SuperShock(double height){ super(height); }
+   public String toString(){
+      return "Super Shock";
+   }
+}
